feat(jit): D-475 C1 — widen JIT table descriptors to u64 (table64 prep)
TableSlice.len/max u32 → u64 (stride 24 → 32; funcptrs 16 → 24, via new
jit_abi.tableslice_{len,funcptrs}_off consts) and JitRuntime.table_size
u32 → u64 (absorbs _pad0 — every subsequent field offset unchanged).
table_grow_fn widens to (delta: u64) → i64 with an overflow-safe add;
jitCallIndirectResolve takes the index at u64. Emitted descriptor
len/table_size loads flip W→X / 32→64-bit on both arches; arm64 bounds
CMPs flip to X-width (idx is W-stage zero-extended, so i32 tables are
outcome-identical); x86_64 keeps .d compares for the i32 path (C3 widens
per idx_type). Prologue pins X25 as the full u64 table-0 size.

Threads per-table idx_types onto ZirFunc (compileWasm validator_tables →
compileOne → func.table_idx_types + tableIdxType accessor) for the C2/C3
width selection. The JitTable64Unsupported guard STAYS until C4.

No behaviour change for i32 tables: test-all green on arm64-macos AND
x86_64-macos (Rosetta), 25539 spec asserts 0 fail.

+++ src/engine/codegen/shared/entry.zig
@@ -195,7 +195,7 @@ pub const EhReifyCtx = jit_abi.EhReifyCtx;
 pub const reifyExnref = jit_abi.reifyExnref;
 pub const SegmentSlice = jit_abi.SegmentSlice;
 pub const TableSlice = jit_abi.TableSlice;
-pub const table_no_max: u32 = jit_abi.table_no_max;
+pub const table_no_max: u64 = jit_abi.table_no_max;
 pub const ElemSlice = jit_abi.ElemSlice;
 pub const TableJitCallInfo = jit_abi.TableJitCallInfo;
