+++ src/engine/runner.zig
@@ -1192,7 +1192,8 @@ pub const JitInstance = struct {
 
     /// D-496 — slot count of table `idx` (full index space, matching
     /// `JitRuntime.tables_ptr`), backing C-API `wasm_table_size`. Null if OOB.
-    pub fn tableLen(self: *JitInstance, idx: u32) ?u32 {
+    /// u64 per D-475 (table64); the u32-shaped wasm-c-api boundary saturates.
+    pub fn tableLen(self: *JitInstance, idx: u32) ?u64 {
         if (idx >= self.owned.rt.tables_count) return null;
         return self.owned.rt.tables_ptr[idx].len;
     }
@@ -1271,7 +1272,7 @@ pub const JitInstance = struct {
     /// fail-safe (the host `init` may be a forged ref, never dereferenced) — a
     /// callable grown funcref slot then needs `wasm_table_set` (mirror-aware). Grow
     /// is bounded by the pre-allocated capacity (a no-max table has no headroom).
-    pub fn growTable(self: *JitInstance, tableidx: u32, init_ref: u64, delta: u32) ?u32 {
+    pub fn growTable(self: *JitInstance, tableidx: u32, init_ref: u64, delta: u32) ?u64 {
         if (tableidx >= self.owned.rt.tables_count) return null;
         const old = setup_mod.jitTableGrowHost(&self.owned.rt, tableidx, init_ref, delta);
         if (old < 0) return null;
